+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/exceptions/ReplyMessageNotFound.kt
@@ -2,5 +2,5 @@ package com.github.insanusmokrassar.TelegramBotAPI.bot.exceptions
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.Response
 
-open class ReplyMessageNotFound(response: Response<*>, message: String?, cause: Throwable?) :
-    RequestException(response, message, cause)
+open class ReplyMessageNotFound(response: Response<*>, plainAnswer: String, message: String?, cause: Throwable?) :
+    RequestException(response, plainAnswer, message, cause)