+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/exceptions/RequestException.kt
@@ -5,15 +5,17 @@ import java.io.IOException
 
 fun newRequestException(
     response: Response<*>,
+    plainAnswer: String,
     message: String? = null,
     cause: Throwable? = null
 ) = when (response.description) {
-    "Bad Request: reply message not found" -> ReplyMessageNotFound(response, message, cause)
-    else -> RequestException(response, message, cause)
+    "Bad Request: reply message not found" -> ReplyMessageNotFound(response, plainAnswer, message, cause)
+    else -> RequestException(response, plainAnswer, message, cause)
 }
 
 open class RequestException internal constructor(
     val response: Response<*>,
+    val plainAnswer: String,
     message: String? = null,
     cause: Throwable? = null
 ) : IOException(